+++ LICENSE
@@ -1,6 +1,6 @@
 The MIT License (MIT)
 
-Copyright (c) 2019 Vasanth V <imvasanthv@gmail.com>
+Copyright (c) 2019 Vasanth V <vasanth@ahey.net>
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
