+++ README.md
@@ -1,8 +1,8 @@
-# Hello
+# Ahey
 
 ## A free, p2p, group video call app for the web. No signups. No downloads. Works in all major browsers.
 
-Hello is built using WebRTC, so all your video, audio & text chat is peer-to-peer. Group video call is achieved using WebRTC mesh. So the quality of the call is inversely proportional to the number of people on the call. The sweet number is somewhere around 6 to 8 people in an average high-speed connection.
+Ahey is built using WebRTC, so all your video, audio & text chat is peer-to-peer. Group video call is achieved using WebRTC mesh. So the quality of the call is inversely proportional to the number of people on the call. The sweet number is somewhere around 6 to 8 people in an average high-speed connection.
 
 ---
 
@@ -13,13 +13,13 @@ Hello is built using WebRTC, so all your video, audio & text chat is peer-to-pee
 
 ### How to Build this app locally
 
-[Fork this repo](https://github.com/vasanthv/hello/fork) and then clone it:
+[Fork this repo](https://github.com/vasanthv/ahey/fork) and then clone it:
 
 ```bash
-git clone https://github.com/<your_name>/hello.git
+git clone https://github.com/<your_name>/ahey.git
 ```
 
-`cd hello` and then install dependencies
+`cd ahey` and then install dependencies
 
 ```bash
 npm install
@@ -31,12 +31,12 @@ Run the app
 npm start
 ```
 
-to start the hello server on port 3000. Your hello instance will be running on http://localhost:3000.
+to start the ahey server on port 3000. Your ahey instance will be running on http://localhost:3000.
 
 ### Contributions
 
-Please refer <a href="https://github.com/vasanthv/hello/blob/master/markdowns/CONTRIBUTIONS.md">CONTRIBUTIONS.md</a> for more info.
+Please refer <a href="https://github.com/vasanthv/ahey/blob/master/markdowns/CONTRIBUTIONS.md">CONTRIBUTIONS.md</a> for more info.
 
 ### LICENSE
 
-<a href="https://github.com/vasanthv/hello/blob/master/LICENSE">MIT License</a>
+<a href="https://github.com/vasanthv/ahey/blob/master/LICENSE">MIT License</a>
