+++ config.js
@@ -1,5 +1,9 @@
 module.exports = {
 	NODE_ENV: process.env.NODE_ENV,
-	PORT: process.env.PORT || 3000,
-	CORS_ORIGIN: "https://hello.vasanthv.me:*,http://localhost:3000*",
+	PORT: process.env.PORT || 824,
+	CORS_ORIGIN: "https://ahey.net:*,http://localhost:824*",
+	STATIC_VIEWS: {
+		"/faq": "Frequently asked questions",
+		"/privacy": "Privacy policy",
+	},
 };
