+++ markdowns/CONTRIBUTIONS.md
@@ -1,16 +1,16 @@
-## Hello Contributions
+## Ahey Contributions
 
 Before opening a PR, please read the following guidelines.
 
 ## Current Contribution Status
 
-Hello is currently **not actively accepting contributions**.
+Ahey is currently **not actively accepting contributions**.
 
-The primary purpose of making the Hello source code available is:
+The primary purpose of making the Ahey source code available is:
 
 - To allow others to use the code for personal projects or to satisfy their curiosity.
 - To ensure code transparency, keeping the app honest about its actions.
 
 ### Support the Project
 
-If you would like to support Hello and its development, consider sponsoring in [Github sponsors](https://github.com/sponsors/vasanthv). It helps in the development and maintenance of the application. Your support is always appreciated!
+If you would like to support Ahey and its development, consider sponsoring in [Github sponsors](https://github.com/sponsors/vasanthv). It helps in the development and maintenance of the application. Your support is always appreciated!
