+++ views/faq.ejs
@@ -1,6 +1,6 @@
 <!DOCTYPE html>
 <html lang="en">
-	<%- include('head'); -%>
+<%- include('head'); -%>
 
 	<body>
 		<div class="wrap">
@@ -27,16 +27,16 @@
 			</div>
 
 			<div class="row">
-				<div class="bold">Can I embed Hello to my website?</div>
+				<div class="bold">Can I embed Ahey to my website?</div>
 				<p>
-					Yes, you can embed any Hello channel to your website using iframe. You can find the
-					<a href="https://github.com/vasanthv/hello/blob/master/markdowns/EMBED.md" target="_blank" class="underline"
-						>instructions here</a
-					>.
+					Yes, you can embed any Ahey channel to your website using iframe. You can find the
+					<a href="https://github.com/vasanthv/ahey/blob/master/markdowns/EMBED.md" target="_blank"
+						class="underline">instructions here</a>.
 				</p>
 			</div>
 
 			<%- include('footer'); -%>
 		</div>
 	</body>
+
 </html>
