+++ views/footer.ejs
@@ -1,13 +1,13 @@
 <br />
 <div class="row small light">
 	<nav>
-		&copy; hello.vasanthv.me &nbsp;
+		&copy; ahey.net &nbsp;
 		<a href="/faq">FAQ</a>
 		&middot;
 		<a href="/privacy">Privacy</a>
 		&middot;
-		<a href="mailto:hello@vasanthv.me" target="_blank">Contact</a>
+		<a href="mailto:vasanth@ahey.net" target="_blank">Contact</a>
 		&middot;
-		<a href="https://github.com/vasanthv/hello" target="_blank">Github</a>
+		<a href="https://github.com/vasanthv/ahey" target="_blank">Github</a>
 	</nav>
 </div>
