+++ views/head.ejs
@@ -1,23 +1,19 @@
 <head>
 	<meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
-	<meta
-		name="viewport"
-		content="width=device-width, height=device-height, initial-scale=1.0, maximum-scale=1.0, user-scalable=0"
-	/>
+	<meta name="viewport"
+		content="width=device-width, height=device-height, initial-scale=1.0, maximum-scale=1.0, user-scalable=0" />
 	<meta name="description" content="A free video chat for the web. No sign ups. No downloads." />
-	<meta
-		name="keywords"
-		content="group video chat, video communication, multiparty video chat, video chat, webrtc, peer to peer, p2p"
-	/>
+	<meta name="keywords"
+		content="group video chat, video communication, multiparty video chat, video chat, webrtc, peer to peer, p2p" />
 	<meta name="apple-mobile-web-app-capable" content="yes" />
 	<meta name="apple-mobile-web-app-status-bar-style" content="default" />
-	<meta name="application-name" content="Hello" />
+	<meta name="application-name" content="Ahey" />
 	<meta name="theme-color" content="#ffffff" />
 
 	<!-- Meta data for link fetching -->
-	<meta property="og:url" content="https://hello.vasanthv.me" />
-	<meta property="og:image" content="https://hello.vasanthv.me/banner.png" />
-	<meta property="og:title" content="Hello" />
+	<meta property="og:url" content="https://ahey.net" />
+	<meta property="og:image" content="https://ahey.net/banner.png" />
+	<meta property="og:title" content="Ahey" />
 	<meta property="og:description" content="A free video chat for the web. No sign ups. No downloads. 100% free." />
 
 	<link rel="apple-touch-icon" sizes="180x180" href="/apple-touch-icon.png" />
@@ -29,13 +25,14 @@
 
 	<link rel="stylesheet" href="/styles.css" />
 
-	<% if(page === "channel") { %>
-	<script defer src="/adapter-latest.js"></script>
-	<script defer src="/socket.io/socket.io.js"></script>
-	<script defer src="/vue.global.prod.js"></script>
-	<script defer src="/socket.js"></script>
-	<script defer src="/script.js"></script>
-	<% } %>
+	<% if(page==="channel" ) { %>
+		<script defer src="/adapter-latest.js"></script>
+		<script defer src="/socket.io/socket.io.js"></script>
+		<script defer src="/vue.global.prod.js"></script>
+		<script defer src="/socket.js"></script>
+		<script defer src="/script.js"></script>
+		<% } %>
 
-	<title>Hello / <%= title %></title>
+			<title>Ahey / <%= title %>
+			</title>
 </head>
