+++ views/index.ejs
@@ -5,7 +5,7 @@
 	<body>
 		<div class="wrap">
 			<br />
-			<div class="row title">Hello</div>
+			<div class="row title">Ahey</div>
 			<div class="row bold">A free video chat for the web.</div>
 			<div class="row">
 				<p>A frictionless peer-to-peer video chat app for the web that can be embedded in any web page.</p>
