+++ views/privacy.ejs
@@ -10,7 +10,7 @@
 
 			<div class="row">
 				<p>
-					At Hello, we prioritize your privacy by collecting only the minimal amount of data necessary to ensure a
+					At Ahey, we prioritize your privacy by collecting only the minimal amount of data necessary to ensure a
 					seamless user experience. Specifically, we collect:
 				</p>
 
@@ -29,10 +29,10 @@
 				<label>Key Points:</label>
 				<ul>
 					<li>All video chats are 100% peer-to-peer, meaning they never touch our servers.</li>
-					<li>Hello is purely peer-to-peer, ensuring your video is not sent to our server.</li>
+					<li>Ahey is purely peer-to-peer, ensuring your video is not sent to our server.</li>
 				</ul>
 
-				<p>The creators of Hello are committed to never using or selling your personal data.</p>
+				<p>The creators of Ahey are committed to never using or selling your personal data.</p>
 			</div>
 
 			<div class="row">
