+++ www/manifest.json
@@ -1,6 +1,6 @@
 {
-	"name": "Hello",
-	"short_name": "Hello",
+	"name": "Ahey",
+	"short_name": "Ahey",
 	"description": "A video chat app for the web.",
 	"icons": [
 		{
