+++ www/robots.txt
@@ -1,3 +1,3 @@
 User-agent: *
 Disallow: 
-Sitemap: https://hello.vasanthv.me/sitemap.xml
+Sitemap: https://ahey.net/sitemap.xml
