+++ www/sitemap.xml
@@ -1,26 +1,23 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<urlset
-      xmlns="http://www.sitemaps.org/schemas/sitemap/0.9"
-      xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-      xsi:schemaLocation="http://www.sitemaps.org/schemas/sitemap/0.9
+<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://www.sitemaps.org/schemas/sitemap/0.9
             http://www.sitemaps.org/schemas/sitemap/0.9/sitemap.xsd">
 
 
-<url>
-  <loc>https://hello.vasanthv.me/</loc>
-  <lastmod>2024-07-06T14:59:42+00:00</lastmod>
-  <priority>1.00</priority>
-</url>
-<url>
-  <loc>https://hello.vasanthv.me/faq</loc>
-  <lastmod>2024-07-06T14:59:42+00:00</lastmod>
-  <priority>0.80</priority>
-</url>
-<url>
-  <loc>https://hello.vasanthv.me/privacy</loc>
-  <lastmod>2024-07-06T14:59:42+00:00</lastmod>
-  <priority>0.80</priority>
-</url>
+  <url>
+    <loc>https://ahey.net/</loc>
+    <lastmod>2024-07-06T14:59:42+00:00</lastmod>
+    <priority>1.00</priority>
+  </url>
+  <url>
+    <loc>https://ahey.net/faq</loc>
+    <lastmod>2024-07-06T14:59:42+00:00</lastmod>
+    <priority>0.80</priority>
+  </url>
+  <url>
+    <loc>https://ahey.net/privacy</loc>
+    <lastmod>2024-07-06T14:59:42+00:00</lastmod>
+    <priority>0.80</priority>
+  </url>
 
 
 </urlset>
